+++ CHANGELOG.md
@@ -6,6 +6,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.60] - 2020-10-07
+### Changed
+- WPF.Utils project
+  - InvokeContentControl
+  - InvokeControl
+  - InvokeProgressBar
+  - InvokeTextBox
+### Added
+- WPF.Utils.Tests project
+  - InvokeContentControlTests
+  - InvokeControlTests
+  - InvokeProgressBarTests
+  - InvokeTextBoxTests
+
 ## [0.1.40] - 2020-10-04
 ### Added
 - WPF.Utils project
+  - InvokeContentControl
+  - InvokeControl
+  - InvokeProgressBar
+  - InvokeTextBox